pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Deep and Efficient Deep Research with ReAct & CodeAct Orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "crawl4ai>=0.6.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-litellm>=0.2.1",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.3",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
